Upgrade RX config for flight control PID storage

+++ src/lib/FlightControl/FlightControl.h
@@ -74,7 +74,7 @@ public:
 
 private:
     bool loadPidParameters();
-    bool loadPidBank(FlightControlPid &rollPid, FlightControlPid &pitchPid, FlightControlPid &yawPid, const float *pid, int count);
+    bool loadPidBank(FlightControlPid &rollPid, FlightControlPid &pitchPid, FlightControlPid &yawPid, const int16_t *pid, int count);
     void loadStickTargets(float &throttle, float &rollAngle, float &pitchAngle, float &yawRate);
 
     FlightControlSensorBackend _sensors;
